lecture-studio/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-dotenv==1.1.1
sqlalchemy==2.0.41
aiosqlite==0.21.0
python-multipart==0.0.20
anthropic==0.55.0
httpx==0.28.1
pydantic==2.11.7
websockets==15.0.1
